The Enclave and TransactionService have swapped names, but not roles. The methods from each have been swapped over so that each class maintains in previous role but with the new name.
