+++ nexus-app/src/main/java/com/github/nexus/service/TransactionServiceImpl.java
@@ -1,9 +1,12 @@
 package com.github.nexus.service;
 
 import com.github.nexus.dao.EncryptedTransactionDAO;
-import com.github.nexus.entity.EncryptedTransaction;
+import com.github.nexus.enclave.keys.model.Key;
+import com.github.nexus.enclave.model.MessageHash;
 
 import javax.transaction.Transactional;
+import java.util.Collection;
+import java.util.Map;
 import java.util.logging.Logger;
 
 @Transactional
@@ -18,15 +21,33 @@ public class TransactionServiceImpl implements TransactionService{
     }
 
     @Override
-    public byte[] send(byte[] from, byte[][] to, byte[] payload){
-        encryptedTransactionDAO.save(new EncryptedTransaction("someValue".getBytes(), "somePayload".getBytes()));
-        return "mykey".getBytes();
+    public boolean delete(final MessageHash hash) {
+        return encryptedTransactionDAO.delete(hash);
     }
 
     @Override
-    public byte[] receive(byte[] key, byte[] to) {
-        LOGGER.info("receive");
-        encryptedTransactionDAO.retrieveAllTransactions();
-        return "payload".getBytes();
+    public Collection<String> retrieveAllForRecipient(final Key recipientPublicKey) {
+        return null;
     }
+
+    @Override
+    public String retrievePayload(final MessageHash hash, final Key intendedRecipient) {
+        return null;
+    }
+
+    @Override
+    public String retrieve(final MessageHash hash, final Key sender) {
+        return null;
+    }
+
+    @Override
+    public MessageHash storePayloadFromOtherNode(final byte[] sealedPayload) {
+        return null;
+    }
+
+    @Override
+    public Map<Key, Map<byte[], byte[]>> encryptPayload(final byte[] message, final Key senderPublicKey, final Collection<Key> recipientPublicKeys) {
+        return null;
+    }
+
 }
