+++ nexus-app/src/test/java/com/github/nexus/enclave/EnclaveImplTest.java
@@ -1,54 +1,42 @@
 package com.github.nexus.enclave;
 
-import com.github.nexus.dao.EncryptedTransactionDAO;
-import com.github.nexus.enclave.keys.model.Key;
 import com.github.nexus.enclave.model.MessageHash;
+import com.github.nexus.service.TransactionService;
 import org.junit.Before;
 import org.junit.Test;
-
-import java.util.Collections;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 public class EnclaveImplTest {
 
-    private EncryptedTransactionDAO dao;
+    private TransactionService txService;
 
     private Enclave enclave;
 
     @Before
     public void setUp(){
-        this.dao = mock(EncryptedTransactionDAO.class);
-        enclave = new EnclaveImpl(dao);
+        this.txService = mock(TransactionService.class);
+        enclave = new EnclaveImpl(txService);
     }
-
 
     @Test
     public void testDelete(){
-        when(dao.delete(any())).thenReturn(true);
+        doReturn(true).when(txService).delete(any(MessageHash.class));
+
         enclave.delete(new MessageHash(new byte[0]));
-        verify(dao, times(1)).delete(any());
+
+        verify(txService).delete(any(MessageHash.class));
     }
 
     @Test
-    public void testRetrieveAllForRecipient(){
-        enclave.retrieveAllForRecipient(new Key(new byte[0]));
+    public void testSend(){
+        enclave.send(new byte[0], new byte[0][0], new byte[0]);
     }
 
     @Test
-    public void testRetrievePayload(){
-        enclave.retrievePayload(new MessageHash(new byte[0]), new Key(new byte[0]));
-    }
-
-    @Test
-    public void testRetrieve(){
-        enclave.retrieve(new MessageHash(new byte[0]), new Key(new byte[0]));
-    }
-
-    @Test
-    public void testStorePayloadFromOtherNode(){
-        enclave.storePayloadFromOtherNode(new byte[0]);
+    public void testReceive(){
+        enclave.receive(new byte[0], new byte[0]);
     }
 
     @Test
@@ -56,8 +44,4 @@ public class EnclaveImplTest {
         enclave.store(new byte[0], new byte[0][0], new byte[0]);
     }
 
-    @Test
-    public void testEncryptPayload(){
-        enclave.encryptPayload(new byte[0], new Key(new byte[0]), Collections.EMPTY_LIST);
-    }
 }
